+++ CHANGELOG.md
@@ -19,6 +19,7 @@ A database migration is required to go from v1.x to this version. See docs.
 
  **Notes**:
  - new dropdown and input styles
+ - new resources/stream endpoint
  - user updates are more granular, don't include entire user payload.
  - `user.id` has been moved to a backend property and all frontend apis now query users by username. Swagger has been updated.
  - removed legacy and deprecated properties